requirements.txt
astroid==2.0.4
certifi==2018.10.15
chardet==3.0.4
Django==2.0.1
django-rest==0.0.1
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
pytz==2018.7
requests==2.20.1
urllib3==1.24.1
wrapt==1.10.11
